Move config/flags to vttablet/common
